data-pipeline/flink/requirements.txt
apache-flink==1.20.0
kafka-python==2.0.2
feedparser==6.0.11
beautifulsoup4==4.13.3
openai==1.61.1
python-dotenv==1.0.1
pydantic==2.10.6
psycopg2-binary==2.9.10
hdfs==2.7.3
elasticsearch==8.17.1
pgvector==0.2.4
tiktoken==0.5.1
